+++ CHANGES.md
@@ -7,6 +7,7 @@
 - #1374 Can't load embedded chat when changing `view_mode` between page reloads
 - #1376 Fixed some alignment issues in the sidebar
 - #1378 Message Delivery Receipts were being sent for carbons and own messages
+- #1353 Fixed Message Delivery Receipts not working because of the message "type" attribute
 
 ## 4.0.6 (2018-12-07)